Added BasePage methods
