+++ autoFrame-framework/src/main/java/framework/core/BasePage.java
@@ -1,0 +1,35 @@
+package framework.core;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.PageFactory;
+import org.openqa.selenium.support.pagefactory.AjaxElementLocatorFactory;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+
+public class BasePage {
+    private static final int TIMEOUT= 5;
+    private static final int POLLING= 10;
+
+    protected WebDriver driver;
+    private WebDriverWait wait;
+
+    public BasePage(WebDriver driver){
+        this.driver  = driver;
+        WebDriverWait wait = new WebDriverWait(driver, TIMEOUT, POLLING);
+        PageFactory.initElements(new AjaxElementLocatorFactory(driver, TIMEOUT), this);
+    }
+
+    protected void waitForElementToAppear(By locator){
+        wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
+    }
+
+    protected void waitForElementToDisappear(By locator){
+        wait.until(ExpectedConditions.invisibilityOfElementLocated(locator));
+    }
+
+    protected void waitForTextToDisappear(By locator, String text){
+        wait.until(ExpectedConditions.textToBe(locator,text));
+    }
+}
