+++ autoFrame-framework/src/main/java/framework/core/BaseTest.java
@@ -1,4 +1,4 @@
-package framework;
+package framework.core;
 
 public class BaseTest {
 }
