+++ autoFrame-page-objects/src/main/java/pages/Homepage.java
@@ -1,4 +1,11 @@
 package pages;
 
+
+import org.openqa.selenium.WebDriver;
+
 public class Homepage {
+
+    public Homepage(WebDriver driver) {
+//        super(driver);
+    }
 }
