+++ autoFrame-tests/src/main/java/auto/tests/HomepageTests.java
@@ -1,0 +1,4 @@
+package auto.tests;
+
+public class HomepageTests {
+}
